restructure the dist output
