+++ package.json
@@ -10,7 +10,7 @@
   ],
   "scripts": {
     "dev": "vite",
-    "build": "rm -rf dist && tsc --p ./tsconfig-build.json && vite build",
+    "build": "rm -rf dist && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "prepublishOnly": "npm run build",
@@ -44,7 +44,8 @@
     "storybook": "^8.0.8",
     "typescript": "^5.4.5",
     "vite": "^5.2.9",
-    "vite-plugin-dts": "^3.8.3"
+    "vite-plugin-dts": "^3.8.3",
+    "vite-tsconfig-paths": "^5.0.1"
   },
   "dependencies": {
     "bulma": "^1.0.2"
