+++ src/components/Button.tsx
@@ -1,5 +1,5 @@
 import { ButtonHTMLAttributes } from "react";
-import { Color, Size } from "../types";
+import { Color, Size } from "../types/shared";
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   color?: Color;
