+++ src/components/Hero.tsx
@@ -22,7 +22,7 @@ export const Hero = () => {
                 <div className="container" style={{ padding: "0 0.5rem" }}>
                   <div className="navbar-brand">
                     <a className="navbar-item">
-                      <img src="/markdown-space-logo.svg" alt="Logo" />
+                      {/* <img src="/markdown-space-logo.svg" alt="Logo" /> */}
                     </a>
                     <span
                       className="navbar-burger burger"
