--- src/types/index.ts
@@ -1,15 +1,0 @@
-export type Color =
-  | "white"
-  | "light"
-  | "dark"
-  | "black"
-  | "text"
-  | "ghost"
-  | "primary"
-  | "link"
-  | "info"
-  | "success"
-  | "warning"
-  | "danger";
-
-export type Size = "small" | "normal" | "medium" | "large";
